requirements.txt
altgraph==0.17.4
ansi2html==1.9.1
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
auto-py-to-exe==2.42.0
Babel==2.14.0
bleach==6.1.0
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.9
django-cors-headers==4.3.1
django-extensions==3.2.3
django-livereload-server==0.5.1
django-plotly-dash==2.2.2
djangorestframework==3.14.0
docutils==0.20.1
dpd_components==0.1.0
Eel==0.16.0
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.0
Flask==3.0.0
fqdn==1.5.1
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivymd==1.1.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
nest-asyncio==1.5.8
numpy==1.26.2
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proxy_tools==0.1.0
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pythonnet==3.0.3
pytz==2023.3.post1
pywebview==4.4.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
QtPy==2.4.1
referencing==0.32.0
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.2.3
terminado==0.18.0
tinycss2==1.2.1
tinydb==4.8.0
tornado==6.4
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
whichcraft==0.6.1
widgetsnbextension==4.0.9
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
